separate pg shorthand
